requirements.txt
gitdb==0.6.4
GitPython==1.0.1
prettytable==0.7.2
requests==2.7.0
smmap==0.9.0
